Large refactor of Markd and Posts Classes. Closes #10
The Posts Class was doing too much for what it was.  The "driving" of
things was refactored into the Markd Class.  This also allowed for
making sure that both unpublished posts did not make their way into
post listings, and that pages of posts would always have the correct
number of POSTS_PER_PAGE on them.