Changes to retrieval of post listings
Post listings were getting sorted alphabetically from what was in the
posts directory.  This commit fixes that problem and also removes some
of the $lastPublished indexing craziness that was going on.

Also POSTS_SORT_ORDER is removed since I hadn't actually worked it into
the logic and I couldn't think of a good enough reason to keep it for
now.  It can always come back some other time.