Minor theme system enhancements
Adds in {{markd_header}} and {{markd_footer}} placeholders that will
eventually be used for themes/plugins to be able to inject items into
the header/footer.

Moves jot form javascript to the footer for best practice.

Adds generator meta tag to header for markd, but this will eventually
need to be worked more into the markd framework and added in with a
filter.